Add support for external SD cards
This commit adds a new preference for enabling volumes to be created on
and imported from an external SD card. The location of this SD card
defaults to /mnt/external1 but it is configurable through the
preferences. External SD card volumes become unavailable when the
preference is switched to disable external SD card support. A future
change could make this dynamic so we gray out volumes if we detect that
the external SD card is not mounted.