Restore activity state for EDVolume{List|Browser}Activity
Prior to this commit we used the cop-out way of handling rotation events
by using 'android:configChanges="orientation|screenSize"' for
EDVolume{List|Browser}Activity. This let the app work fine when the
device is rotated, but resulted in the activity not getting destroyed
and recreated, which among other things prevented the ActionBar layout
from being recreated. Especially for tablets such as Nexus 7 the
landscape layout is dramatically more spacious than the portrait layout,
so recreating view layout is desirable.

This change reworks EDVolume{List|Browser}Activity to properly save and
restore activity state so the configChanges workaround is no longer
needed.

The change also contains cleanups in file organization.