Setting and entering PINs per volume
Work in progress...

Done so far: once a volume's key is cached for the first time, the user can set
a PIN. During unlocking, the cached key is used only if the correct PIN has
been supplied.

Needs a database update, so all existing volumes must be deleted and re-added
manually (I think).